WebexEyes/backend/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.2
future==1.0.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
tldextract==5.1.1
tzdata==2024.1
urllib3==1.26.6
Werkzeug==3.0.1
zipp==3.18.1
